+++ jirkov.vladimir/T3/polygon.hpp
@@ -1,0 +1,18 @@
+#ifndef POLYGON_HPP
+#define POLYGON_HPP
+
+#include "point.hpp"
+#include <vector>
+
+namespace jirkov
+{
+  struct Polygon
+  {
+    std::vector< Point > points;
+  };
+
+  std::istream& operator>>(std::istream& in, Polygon& value);
+}
+
+#endif
+
